third_party/blink/web_tests/http/tests/security/subresourceIntegrity/subresource-integrity-script-no-cors-bad-integrity.html

<!DOCTYPE html>
<head>
    <title>SRI on cross origin CORS disabled script, with bad integrity</title>
</head>
<body>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script>
        setup({single_test: true});

        var result = '';
        var scriptLoaded = function() {
            assert_unreached('Script loaded.');
            done();
        };
        var scriptError = function() {
            assert_equals(result, '', 'Script did not run.');
            done();
        };
    </script>
    <script crossorigin integrity="sha256-deadbeef" src="http://localhost:8000/security/resources/cors-script.php?cors=false&value=ran2" onload="scriptLoaded();" onerror="scriptError();"></script>
</body>
</html>
